refactor hook and add docstring
